Attempting to optimize breaks
Broke deployment when attempting to optimize.

+++ governance/Democratized.sol
@@ -48,29 +48,30 @@ contract Democratized is ERC721Holder, ERC1155Holder {
     uint propositionID;
   }
 
-  mapping (uint => ETHRequest) ethRequests;
-  mapping (uint => ERC20Request) erc20Requests;
-  mapping (uint => ERC721Request) erc721Requests;
-  mapping (uint => ERC1155Request) erc1155Requests;
-  mapping (uint => ERC1155BatchRequest) erc1155BatchRequests;
+  // We don't need 2^256 potential requests and using 32 might help with tight packing
+  mapping (uint32 => ETHRequest) ethRequests;
+  mapping (uint32 => ERC20Request) erc20Requests;
+  mapping (uint32 => ERC721Request) erc721Requests;
+  mapping (uint32 => ERC1155Request) erc1155Requests;
+  mapping (uint32 => ERC1155BatchRequest) erc1155BatchRequests;
 
-  uint ethRequestCount;
-  uint erc20RequestCount;
-  uint erc721RequestCount;
-  uint erc1155RequestCount;
-  uint erc1155BatchRequestCount;
+  uint32 ethRequestCount;
+  uint32 erc20RequestCount;
+  uint32 erc721RequestCount;
+  uint32 erc1155RequestCount;
+  uint32 erc1155BatchRequestCount;
 
 
   /// @dev Not sure if additional or less information should be made available. Include sender?
   event ETHWithdrawRequested(uint amt, uint requestID);
-  event ERC20WithdrawRequested(address token, uint amt, uint requestID);
-  event ERC721WithdrawRequested(address token, uint tokeNID, uint requestID);
-  event ERC1155WithdrawRequested(address token, uint tokenID, uint amt, uint requestID);
-  event ERC1155WithdrawRequested(address token, uint[] tokenIDs, uint[] amts, uint requestID);
-  event ERC20WithdrawExecuted(address token, uint amt, uint requestID);
-  event ERC721WithdrawExecuted(address token, uint tokenID, uint requestID);
-  event ERC1155WithdrawExecuted(address token, uint tokenID, uint amt, uint requestID);
-  event ERC1155WithdrawExecuted(address token, uint[] tokenIDs, uint[] amts, uint requestID);
+  event ERC20WithdrawRequested(address token, uint amt, uint32 requestID);
+  event ERC721WithdrawRequested(address token, uint tokeNID, uint32 requestID);
+  event ERC1155WithdrawRequested(address token, uint tokenID, uint amt, uint32 requestID);
+  event ERC1155WithdrawRequested(address token, uint[] tokenIDs, uint[] amts, uint32 requestID);
+  event ERC20WithdrawExecuted(address token, uint amt, uint64 requestID);
+  event ERC721WithdrawExecuted(address token, uint tokenID, uint32 requestID);
+  event ERC1155WithdrawExecuted(address token, uint tokenID, uint amt, uint32 requestID);
+  event ERC1155WithdrawExecuted(address token, uint[] tokenIDs, uint[] amts, uint32 requestID);
   
   constructor(VotingMachine machine_) {
     voting = machine_;
@@ -79,7 +80,7 @@ contract Democratized is ERC721Holder, ERC1155Holder {
     return address(voting);
   }
   function _requestETHWithdraw(uint amt, uint threshold, address payable receiver, uint startTime, uint endTime) internal returns (uint propositionID) {
-    uint requestID = ethRequestCount;
+    uint32 requestID = ethRequestCount;
     ethRequestCount++;
     uint propID = voting.addProposition(msg.sender, threshold, startTime, endTime);
     ethRequests[requestID] = ETHRequest(amt, receiver, propID);
@@ -87,7 +88,7 @@ contract Democratized is ERC721Holder, ERC1155Holder {
     return propID;
   }
   function _requestWithdraw20(address token, uint amt, uint threshold, address receiver, uint startTime, uint endTime) internal returns (uint propositionID) {
-    uint requestID = erc20RequestCount;
+    uint32 requestID = erc20RequestCount;
     erc20RequestCount++;
     uint propID = voting.addProposition(msg.sender, threshold, startTime, endTime);
     erc20Requests[requestID] = ERC20Request(token, amt, receiver, propID);
@@ -95,21 +96,21 @@ contract Democratized is ERC721Holder, ERC1155Holder {
     return propID;
   }
   function _requestWithdraw721(address token, uint tokenID, uint threshold, address receiver, uint startTime, uint endTime) internal returns (uint propositionID) {
-    uint requestID = erc721RequestCount;
+    uint32 requestID = erc721RequestCount;
     erc721RequestCount++;
     uint propID = voting.addProposition(msg.sender, threshold, startTime, endTime);
     erc721Requests[requestID] = ERC721Request(token, tokenID, receiver, propID);
     return propID;
   }
   function _requestWithdraw1155(address token, uint tokenID, uint amt, uint threshold, address receiver, uint startTime, uint endTime) internal returns (uint propositionID) {
-    uint requestID = erc1155RequestCount;
+    uint32 requestID = erc1155RequestCount;
     erc1155RequestCount++;
     uint propID = voting.addProposition(msg.sender, threshold, startTime, endTime);
     erc1155Requests[requestID] = ERC1155Request(token, tokenID, amt, receiver, propID);
     return propID;
   }
   function _requestWithdraw1155(address token, uint[] memory tokenIDs, uint[] memory amts, uint threshold, address receiver, uint startTime, uint endTime) internal returns (uint propositionID) {
-    uint requestID = erc1155BatchRequestCount;
+    uint32 requestID = erc1155BatchRequestCount;
     erc1155BatchRequestCount++;
     uint propID = voting.addProposition(msg.sender, threshold, startTime, endTime);
     erc1155BatchRequests[requestID] = ERC1155BatchRequest(token, tokenIDs, amts, receiver, propID);
@@ -131,28 +132,28 @@ contract Democratized is ERC721Holder, ERC1155Holder {
   function requestWithdrawERC1155(address token, uint[] memory tokenIDs, uint[] memory amts, address receiver) public virtual returns (uint propositionID) {
     return _requestWithdraw1155(token, tokenIDs, amts, 5000000, receiver, block.timestamp + 24 hours , block.timestamp + 48 hours);
   }
-  function executeETHWithdraw(uint requestID) public {
+  function executeETHWithdraw(uint32 requestID) public {
     require(requestID < ethRequestCount, "No such request.");
     ETHRequest memory request = ethRequests[requestID];
     /// @dev Make sure proposition is set to executed FIRST
     voting.executeProposition(request.propositionID, msg.sender);
     request.receiver.transfer(request.amt);
   }
-  function executeWithdrawERC20(uint requestID) public {
+  function executeWithdrawERC20(uint32 requestID) public {
     require(requestID < erc20RequestCount, "No such request.");
     ERC20Request memory request = erc20Requests[requestID];
     voting.executeProposition(request.propositionID, msg.sender);
     ERC20 token = ERC20(request.token);
     token.transferFrom(address(this), request.receiver, request.amt);
   }
-  function executeWithdrawERC721(uint requestID) public {
+  function executeWithdrawERC721(uint32 requestID) public {
     require(requestID < erc721RequestCount, "No such request.");
     ERC721Request memory request = erc721Requests[requestID];
     voting.executeProposition(request.propositionID, msg.sender);
     ERC721 token = ERC721(request.token);
     token.safeTransferFrom(address(this), request.receiver, request.tokenID);
   }
-  function executeWithdrawERC1155(uint requestID) public {
+  function executeWithdrawERC1155(uint32 requestID) public {
     require(requestID < erc1155RequestCount, "No such request.");
     ERC1155Request memory request = erc1155Requests[requestID];
     voting.executeProposition(request.propositionID, msg.sender);
@@ -160,7 +161,7 @@ contract Democratized is ERC721Holder, ERC1155Holder {
     /// @dev I'm not sure if I should include the data parameter or not.
     token.safeTransferFrom(address(this), request.receiver, request.tokenID, request.amt, "");
   }  
-  function executeWithdrawERC1155Batch(uint requestID) public {
+  function executeWithdrawERC1155Batch(uint32 requestID) public {
     require(requestID < erc1155BatchRequestCount, "No such request.");
     ERC1155BatchRequest memory request = erc1155BatchRequests[requestID];
     voting.executeProposition(request.propositionID, msg.sender);
